refactoring: find methods from user repository, packages
