+++ user-microservice/src/test/java/com/wasd/usermicroservice/service/impl/UserServiceImplTest.java
@@ -4,8 +4,8 @@ import com.wasd.usermicroservice.data.user.UserRequest;
 import com.wasd.usermicroservice.data.user.UserResponse;
 import com.wasd.usermicroservice.exception.UserAlreadyExistsException;
 import com.wasd.usermicroservice.exception.UserNotFoundException;
-import com.wasd.usermicroservice.persistence.model.User;
-import com.wasd.usermicroservice.persistence.repository.UserRepository;
+import com.wasd.usermicroservice.persistence.user.User;
+import com.wasd.usermicroservice.persistence.user.UserRepository;
 import com.wasd.usermicroservice.service.SecurityService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
